pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "twig"
version = "1.5"
description = "An evolution of Leviathan - flexible media to LLM context converter with RAG capabilities"
requires-python = ">=3.12,<3.13.4"
dependencies = [
    "rich",
    "pillow",
    "Chain",
]

[project.scripts]
twig = "twig.twig:main"  # src/twig/twig.py -> def main()

# ── Hatchling (build) ──────────────────────────────────────────────────────────
# Point Hatchling at the src/ package; this is the critical bit for src-layout.
[tool.hatch.build.targets.wheel]
packages = ["src/twig"]
sources = ["src"]

# If you ship non-.py files inside the package (configs, prompts, assets),
# list them so they’re included in wheels and sdists.
[tool.hatch.build]
include = [
  "src/twig/**/*.txt",
  "src/twig/**/*.md",
  "src/twig/**/*.json",
  "src/twig/**/*.yaml",
  "src/twig/**/*.yml",
  "src/twig/**/*.ini",
  "src/twig/**/*.jinja*",
  "src/twig/assets/**",
]

[tool.hatch.build.targets.sdist]
include = ["src/**", "README.*", "LICENSE*"]

# ── Pytest ─────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
addopts = "-v -s --tb=short --no-header --showlocals --pdb -x"
log_cli = true
log_cli_level = "INFO"

# ── Ruff ───────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 88
target-version = "py312"  # align with requires-python

[tool.ruff.lint]
select = [
    "F", "E",
    "UP",
    "SIM",
    "ARG",
    "B",
    "RUF",
    "S102", "S103", "S108",
    "PERF",
    "FBT003",
]
ignore = [
    "SIM118",
    "ARG001", "ARG002",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG", "FBT"]
"scripts/**/*.py" = ["ARG", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["twig"]
force-sort-within-sections = true
split-on-trailing-comma = true

# ── uv sources (local/editable deps) ───────────────────────────────────────────
[tool.uv.sources]
database-clients = { path = "../database-clients", editable = true }
Chain = { path = "../Chain", editable = true }

[tool.uv.workspace]
members = [
    ".",
]


